requirements.txt
aiogram==3.10.0
python-dotenv==1.0.0
pillow==10.1.0
sqlalchemy==2.0.23
aiosqlite==0.19.0